Fix CNAME resolution
Stripping the trailing . will cause DNS to interpret the CNAME record as
a relative domain name. The resolver will then attempt to append a
locally configured search domain to the target. This is fine when the
network (or local resolver) is set up to handle search domains, but it
can't be depended on.

This fix allows DNS to intepret the CNAME as an FQDN by _not_ stripping
the ".". This, technically speaking, is a breaking change if any users
had their configuration set at "cname a.b.c." when they expected a
relative domain name to be configured. The fix for those users would be
to simply edit their configuration to read "cnam a.b.c".